osm/josm: make sure stops are accessible by all their refs
In some GTFS datasets, stops may be duplicated, in the sense that some
stops have same name and same coordinates, but different "ref" (usually
called stop_id in GTFS terms).

In OpenStreetMap, these stops correspond to only one node, and stop_id
are separated by a semicolon, so we might see something like:

    ref=12345;67890

In the export_routes function, only the first ref was kept, leading to
assertion errors down the road. This commit makes sure a stop can be
accessed by all its references, not on the first one.