gtfs/importer: move out specific STM logic
A goal of this project is to be able to support multiple GTFS providers,
at least those of the greater Montréal area. Currently, the GTFS
importer code is very STM-oriented (STM being the transport agency of
the Montréal Island).

This commit is a first step to decouple the main import logic from
specific dataset.