overpass: add capability to generate the cache file
This application has been capable for some time to use a cache file
instead of doing requests to the Overpass API, but it was not able to
generate it, relying on someone creating the cache file manually. That
made the feature hard to discover and to use.

gtfsimporter gains a "generate-cache" command. Basically, it queries the
Overpass API and stores the result in a file. Further commands can use
this cache by adding `--osm-cache <file>` on the command line.