osm/josm: let export_route functions indicate success
gtfsimporter uses argparse to parse arguments. One of the option when
exporting a route is `--dest <output file>`, which is defined of being
of type=argparse.FileType('w'). The main drawback of that is that the
file is created unconditionnally.

So for instance, when trying to export a route and the export fails for
some reasons, an empty file is created. To prevent that, the first step
is to be able to detect if exporting a route is successful or not.