gtfs/importer: don't skip first row of stop_times.txt data
This next() was originally used because the csv was not parsed using a
DictReader, so the first row had to be skipped because it was the header
row. But now, this has the consequence of making load_stop_times miss
the first entry of stop_times.txt.

(Fixes: "4a011338329: gtfs/importer: move out specific STM logic")