major overhaul: ugly big commit
After a big period of inactivity, I got back to this project in order to
be able to conflate more easily existing OSM data with new GTFS data.

The main issue was that route could only be exported, but not conflated
with existing data. As I had no real idea of how I would tackle that, I
started modifying the code all over the place to end up with something
that works-ish.

We are at this stage, with one ugly commit to rule it all. Now is time
to get back to atomic and clean commits.