cli/loader: add new classes
The goal is to split the `main.py` file in more manageable units. This
starts by creating a `cli` module, to contain all helpers, classes,
functions related to the command line.

The first helpers to appear are loaders. This class can parse arguments
from the command line to load GTFS and OSM schedules. This is
boilerplate that's needed everywhere (we need data to work on right?),
hopefully this will reduce the clutter a bit.