cli/cache: add class to generate caches
This class is in charge of generating caches for GTFS and OSM schedules.
For instance, loading GTFS schedule for the STM takes up to 8min on my
machine (a lot of time is spent deduplicating trips). Once pickled, the
schedule takes only a few milliseconds to load, making development and
tests way more manageable.