cli/stop: add export-missing subcommands
This command only export stops that are present in GTFS and missing in
OSM, based on their "ref".

WARNING: Note that it doesn't mean that the stops don't exist in OSM,
only that there are no stops with the same ref. Stops without the "ref"
tags, so I might be useful to combine this step with the conflation
plugin in JOSM.