osm/trip: add capability to mark import as failed
Importing trips from OSM might fail because nodes are not in the
right order, have invalid roles, or have ways with roles. This is not
something that can be solved manually and require manual intervention.

But we are still interested in knowing that these relations exist,
because we don't want to re-export them for instance. So add a way to
mark this trip as being failed and non-modifiable.