osm/elements: OsmTrip: raise error when modifying in-error trips
Some trips' import might fail because we're not able to parse their
contents properly. The current version only set set_tag to None, which
creates error of the like "NonType object is not callable". But that's
not very helpful.

Instead, replace set_tag by another one that raises an exception. That
let the caller get a better understanding of what the issue might be.