conflation/routes: add function get matching OSM routes
Routes in OSM might be only partially described, eg. only the ref is
there, but not the operator or network tags. It's quite impossible to
try to match these routes accurately if they only have partial
information, so introduce a helper function that could be used by the
CLI module to find the right route.